+++ 2023/AdventOfCode2023/AdventOfCode2023/Day03.cs
@@ -1,4 +1,5 @@
-﻿using AdventOfCode2023_1.Shared;
+﻿using AdventOfCode2023_1.Models.Day03;
+using AdventOfCode2023_1.Shared;
 
 namespace AdventOfCode2023_1;
 
@@ -183,28 +184,4 @@ public class Day03: DayBase
         }
     }
     # endregion
-}
-
-internal class EngineNumber(int rowIndex, int columnIndex, int numberLength, int number)
-{
-    public int RowIndex { get; } = rowIndex;
-    public int ColumnIndex { get; } = columnIndex;
-    public int NumberLength { get; set; } = numberLength;
-    public int Number { get; set; } = number;
-    public bool IsPartNumber { get; set; }
-}
-
-internal class EngineSymbol(int rowIndex, int columnIndex, char symbol)
-{
-    public int RowIndex { get; } = rowIndex;
-    public int ColumnIndex { get; } = columnIndex;
-    public string Symbol { get; } = symbol.ToString();
-    public bool IsGear { get; set; }
-    public List<EngineNumber>? AdjacentPartNumbers { get; set; }
-    public int GearRatio { get; set; }
-
-    internal int GetGearRatio()
-    {
-        return AdjacentPartNumbers?.Aggregate(1, (current, partNumber) => current * partNumber.Number) ?? 0;
-    }
 }
