+++ 2023/AdventOfCode2023/AdventOfCode2023/DayBase.cs
